cfa_server/requirements.txt
asgiref==3.7.2
# backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
geographiclib==2.0
geopy==2.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
django-filter==23.2
django-debug-toolbar==4.1.0
crispy-bootstrap5==0.7
pyotp==2.6.0
pytesseract==0.3.10
opencv-python==4.8.0.74
